refactor: isolate MCP and Pi adapters
