+++ src/adapters/mcp/register-prompts.ts
